refactor: #554 응답의 소수 데이터를 Float 타입으로 변경 (#559)
* refactor: 응답의 소수 데이터를 Float 타입으로 변경

* style: 개행 수정

+++ backend/ddang/src/main/java/com/ddang/ddang/review/presentation/dto/request/CreateReviewRequest.java
@@ -17,6 +17,6 @@ public record CreateReviewRequest(
 
         @NotNull(message = "점수가 입력되지 않았습니다.")
         @Positive(message = "점수는 양수여야 합니다.")
-        Double score
+        Float score
 ) {
 }
